pyproject.toml
[tool.poetry]
name = "django-and-react"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "django_and_react"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^10.0.0"
pyjwt = "^2.8.0"
psycopg = "^3.1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
